myocrproject/ocrmodel/OCR/requirements.txt
shapely
scikit-image
imgaug
pyclipper
lmdb
tqdm
numpy
visualdl
rapidfuzz
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66
cython
lxml
premailer
openpyxl
attrdict
PyMuPDF<1.21.0
paddlepaddle
onnxruntime
openai
beautifulsoup4
mysql-connector-python
Polygon3
lanms-neo==1.0.2